pyproject.toml
[tool.poetry]
name = "gcalcli"
version = "4.3.0"
description = "Google Calendar Command Line Interface"
authors = ["Eric Davis, Brian Hartvigsen, Joshua Crowgey <email@email.com>"]
readme = "README.md"
license = "Proprietary"
homepage = "https://github.com/insanum/gcalcli"
repository = "https://github.com/insanum/gcalcli"
include = ["ChangeLog", "README.md"]

[tool.poetry.urls]
changelog = "https://github.com/insanum/gcalcli/blob/master/ChangeLog"

[tool.poetry.dependencies]
python = "^3.8"
google-api-python-client = ">=1.4"
httplib2 = "*"
oauth2client = "*"
parsedatetime = "*"
python-dateutil = "*"
durations-nlp = "^1.0.1"

[tool.poetry.scripts]
gcalcli = "gcalcli.cli:main"

[tool.poetry.group.dev.dependencies]
mypy = "^1.11.1"
google-api-python-client-stubs = "^1.27.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

## Exclude a variety of commonly ignored directories.
[tool.ruff]
unsafe-fixes = true

[tool.ruff.format]
preview = true
